hsi-12409 [BE] Crear caso de uso para poblar el esquema de medicacion comercial

+++ back-end/interoperability/src/main/java/net/pladema/medication/domain/commercialMedicationArticle/NewDrug.java
@@ -9,6 +9,8 @@ import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -20,7 +22,7 @@ public class NewDrug {
 	private Integer id;
 
 	@XmlElement(name = "po")
-	private Integer potency;
+	private BigDecimal potency;
 
 	@XmlElement(name = "upo")
 	private Integer potencyUnityId;
