+++ back-end/interoperability/src/main/java/net/pladema/medication/domain/commercialMedicationArticle/VademecumAndCoverage.java
@@ -9,6 +9,8 @@ import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -17,16 +19,16 @@ import javax.xml.bind.annotation.XmlRootElement;
 public class VademecumAndCoverage {
 
 	@XmlElement(name = "c1")
-	private Float c1;
+	private BigDecimal c1;
 
 	@XmlElement(name = "c2")
-	private Float c2;
+	private BigDecimal c2;
 
 	@XmlElement(name = "c3")
 	private String c3;
 
 	@XmlElement(name = "c4")
-	private Float c4;
+	private BigDecimal c4;
 
 	@XmlElement(name = "c5")
 	private String c5;
